Fix the first day of week logic (mac)
