+++ src/ui/osx/TogglDesktop/Feature/Calendar/CalendarViewController.swift
@@ -8,11 +8,59 @@
 
 import Cocoa
 
-class CalendarViewController: NSViewController {
+final class CalendarViewController: NSViewController {
+
+    private struct Constants {
+
+        static let cellID = NSUserInterfaceItemIdentifier("DateCell")
+        static let cellNibName = NSNib.Name("DateCell")
+    }
+
+    // MARK: OUTLET
+
+    @IBOutlet weak var collectionView: NSCollectionView!
+
+    // MARK: Variables
+
+    fileprivate lazy var dataSource: CalendarDataSource = CalendarDataSource(selectedDate)
+    private var selectedDate = Date() {
+        didSet {
+            dataSource.selectedDate = selectedDate
+        }
+    }
+
+    // MARK: View Cycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do view setup here.
+        initCommon()
+        initCollectionView()
     }
-    
+
+    // MARK: Public
+
+    func prepareLayout(with date: Date) {
+        selectedDate = date
+    }
+}
+
+// MARK: Private
+
+extension CalendarViewController {
+
+    fileprivate func initCommon() {
+
+    }
+
+    fileprivate func initCollectionView() {
+        collectionView.register(NSNib(nibNamed: Constants.cellNibName, bundle: nil),
+                                forItemWithIdentifier: Constants.cellID)
+        collectionView.dataSource = dataSource
+        collectionView.delegate = dataSource
+        let flow = NSCollectionViewFlowLayout()
+        flow.itemSize = CGSize(width: 32, height: 26)
+        flow.minimumLineSpacing = 20
+        flow.minimumInteritemSpacing = 2
+        collectionView.collectionViewLayout = flow
+    }
 }
